WIP refactor to move button
