+++ src/pages/ReportPayouts/OutcomesTable.tsx
@@ -1,23 +1,15 @@
-import { BigNumber, formatUnits } from 'ethers/utils'
+import { BigNumber } from 'ethers/utils'
 import React, { useEffect, useState } from 'react'
-import { Controller, useForm } from 'react-hook-form'
+import { Controller, FormContextValues } from 'react-hook-form'
 import styled from 'styled-components'
 
-import { Button } from 'components/buttons/Button'
 import { BigNumberInputWrapper } from 'components/form/BigNumberInputWrapper'
-import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
-import { Error, ErrorContainer } from 'components/pureStyledComponents/Error'
 import { TableWrapper } from 'components/pureStyledComponents/TableWrapper'
-import { FullLoading } from 'components/statusInfo/FullLoading'
-import { IconTypes } from 'components/statusInfo/common'
 import { ZERO_BN } from 'config/constants'
-import { useConditionContext } from 'contexts/ConditionContext'
-import { Web3ContextStatus, useWeb3ConnectedOrInfura } from 'contexts/Web3Context'
 import { useQuestion } from 'hooks/useQuestion'
+import { FormInputs } from 'pages/ReportPayouts/Contents'
 import { GetCondition_condition } from 'types/generatedGQL'
-import { getLogger } from 'util/logger'
 import { divBN } from 'util/tools'
-import { Status } from 'util/types'
 
 const Wrapper = styled.form``
 
@@ -96,6 +88,8 @@ TH.defaultProps = {
 
 interface Props {
   condition: GetCondition_condition
+  formMethods: FormContextValues<FormInputs>
+  decimals: number
 }
 
 interface Outcome {
@@ -103,44 +97,12 @@ interface Outcome {
   probability: number
 }
 
-interface FormInputs {
-  payouts: BigNumber[]
-}
-
-const logger = getLogger('OutcomSlotsToReport')
-
-const ORACLE_NOT_VALID_TO_REPORT_ERROR = 'The connected user is a not allowed to report payouts'
-const PAYOUTS_POSITIVE_ERROR = 'At least one payout must be positive'
-
-const DECIMALS = 2
-
-export const OutcomesTable = ({ condition }: Props) => {
-  const { _type: status, CTService, address, connect } = useWeb3ConnectedOrInfura()
-  const { clearCondition } = useConditionContext()
-  const { oracle, outcomeSlotCount, questionId } = condition
+export const OutcomesTable = ({ condition, decimals, formMethods }: Props) => {
+  const { outcomeSlotCount, questionId } = condition
   const { outcomesPrettier } = useQuestion(questionId, outcomeSlotCount)
   const [outcomes, setOutcomes] = useState<Outcome[]>([])
-  const [payoutEmptyError, setPayoutEmptyError] = useState(false)
-  const [transactionStatus, setTransactionStatus] = useState<Maybe<Status>>(null)
-  const [oracleNotValidError, setOracleNotValidError] = useState(true)
-  const [error, setError] = useState<Maybe<Error>>(null)
 
-  const {
-    control,
-    formState: { dirty },
-    getValues,
-    handleSubmit,
-    watch,
-  } = useForm<FormInputs>({ mode: 'onSubmit' })
-
-  // Check if the sender is valid
-  useEffect(() => {
-    if (status === Web3ContextStatus.Connected && address) {
-      setOracleNotValidError(oracle.toLowerCase() !== address.toLowerCase())
-    } else {
-      setOracleNotValidError(true)
-    }
-  }, [status, address, oracle])
+  const { control, getValues } = formMethods
 
   useEffect(() => {
     let cancelled = false
@@ -157,16 +119,6 @@ export const OutcomesTable = ({ condition }: Props) => {
       cancelled = true
     }
   }, [outcomesPrettier, outcomes.length])
-
-  const watchPayouts = watch('payouts')
-
-  // Validate payouts (positive, at least one non 0)
-  useEffect(() => {
-    if (watchPayouts && watchPayouts.length > 0 && dirty) {
-      const nonZero = (currentValue: BigNumber) => !currentValue.isZero()
-      setPayoutEmptyError(!watchPayouts.some(nonZero))
-    }
-  }, [watchPayouts, dirty])
 
   const onChange = (value: BigNumber, index: number) => {
     const values = Object.values(getValues())
@@ -193,42 +145,8 @@ export const OutcomesTable = ({ condition }: Props) => {
     setOutcomes(outcomesValues)
   }
 
-  const onSubmit = async (data: FormInputs) => {
-    // Validate exist at least one payout
-    const { payouts } = data
-    try {
-      if (status === Web3ContextStatus.Connected) {
-        setTransactionStatus(Status.Loading)
-
-        const payoutsNumbered = payouts.map((payout: BigNumber) =>
-          Number(formatUnits(payout, DECIMALS))
-        )
-        await CTService.reportPayouts(questionId, payoutsNumbered)
-
-        setTransactionStatus(Status.Ready)
-
-        // Setting the condition to '', update the state of the provider and reload the HOC component, works like a reload
-        clearCondition()
-      } else if (status === Web3ContextStatus.Infura) {
-        connect()
-      }
-    } catch (err) {
-      setError(err)
-      logger.error(err)
-    } finally {
-      setTransactionStatus(Status.Ready)
-    }
-  }
-
-  // Variable used to disable the submit button, check for payouts not empty and the oracle must be valid
-  const disableSubmit =
-    !dirty ||
-    payoutEmptyError ||
-    (status === Web3ContextStatus.Connected && oracleNotValidError) ||
-    transactionStatus === Status.Loading
-
   return (
-    <Wrapper onSubmit={handleSubmit(onSubmit)}>
+    <Wrapper>
       <TableWrapper>
         <Table>
           <THead>
@@ -249,7 +167,7 @@ export const OutcomesTable = ({ condition }: Props) => {
                     <Controller
                       as={Textfield}
                       control={control}
-                      decimals={DECIMALS}
+                      decimals={decimals}
                       defaultValue={new BigNumber(0)}
                       name={`payouts[${index}]`}
                       onChange={(value) => {
@@ -265,27 +183,6 @@ export const OutcomesTable = ({ condition }: Props) => {
           </TBody>
         </Table>
       </TableWrapper>
-      {transactionStatus === Status.Loading && (
-        <FullLoading
-          actionButton={
-            error ? { text: 'OK', onClick: () => setTransactionStatus(Status.Ready) } : undefined
-          }
-          icon={error ? IconTypes.error : IconTypes.spinner}
-          message={error ? error.message : 'Waiting...'}
-          title={error ? 'Error' : 'Report payout'}
-        />
-      )}
-      <ErrorContainer>
-        {payoutEmptyError && <Error>{PAYOUTS_POSITIVE_ERROR}</Error>}
-        {status === Web3ContextStatus.Connected && oracleNotValidError && (
-          <Error>{ORACLE_NOT_VALID_TO_REPORT_ERROR}</Error>
-        )}
-      </ErrorContainer>
-      <ButtonContainer>
-        <Button disabled={disableSubmit} type="submit">
-          Report
-        </Button>
-      </ButtonContainer>
     </Wrapper>
   )
 }
